Split new_spec into multiple specs, based on flags given
